Deployment/requirements.txt
torch
ultralytics
transformers
streamlit
opencv-python
numpy
Pillow
huggingface-hub
verovio
tiktoken
streamlit-webrtc
python-dotenv
